+++ test_conformance/basic/test_multireadimageonefmt.c
@@ -109,7 +109,7 @@ int test_multireadimageonefmt(cl_device_id device, cl_context context, cl_comman
 	int img_height = 512;
 	int i, err;
     size_t origin[3] = {0, 0, 0};
-    size_t region[3] = {img_width, img_height, 1};
+    size_t region[3] = {static_cast<size_t>(img_width), static_cast<size_t>(img_height), 1};
     size_t length = img_width * img_height * 4 * sizeof(float);
     MTdata d;
