+++ test_conformance/basic/test_readimage_int16.c
@@ -88,7 +88,7 @@ int test_readimage_int16(cl_device_id device, cl_context context, cl_command_que
 	int img_height = 512;
 	int err;
     size_t origin[3] = {0, 0, 0};
-    size_t region[3] = {img_width, img_height, 1};
+    size_t region[3] = {static_cast<size_t>(img_width), static_cast<size_t>(img_height), 1};
     size_t length = img_width * img_height * 4 * sizeof(cl_ushort);
     MTdata d;
