+++ test_conformance/profiling/execute.c
@@ -338,7 +338,7 @@ static int kernelFilter( cl_device_id device, cl_context context, cl_command_que
 	
 	// read output image
 	size_t origin[3] = { 0, 0, 0 };
-	size_t region[3] = { w, h, 1 };
+	size_t region[3] = { static_cast<size_t>(w), static_cast<size_t>(h), 1 };
 	err = clEnqueueReadImage( queue, memobjs[1], true, origin, region, 0, 0, outptr, 0, NULL, NULL); 
 	if( err != CL_SUCCESS ){ 
 		print_error( err, "clReadImage failed\n" );
